automation/requirements.txt
facebook-wda
tidevice
